+++ elide-core/src/main/java/com/yahoo/elide/Elide.java
@@ -32,6 +32,7 @@ import com.yahoo.elide.parsers.JsonApiParser;
 import com.yahoo.elide.parsers.PatchVisitor;
 import com.yahoo.elide.parsers.PostVisitor;
 import com.yahoo.elide.security.User;
+import com.yahoo.elide.utils.coerce.CoerceUtil;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.antlr.v4.runtime.misc.ParseCancellationException;
@@ -67,6 +68,10 @@ public class Elide {
         this.dataStore = elideSettings.getDataStore();
         this.dataStore.populateEntityDictionary(elideSettings.getDictionary());
         this.mapper = elideSettings.getMapper();
+
+        elideSettings.getSerdes().forEach((targetType, serde) -> {
+            CoerceUtil.register(targetType, serde);
+        });
     }
 
     /**